// src/app/pages/students-list/StudentDetails.interface.ts
export interface StudentDetailsInterface {
  id?: string;
  studentName: string;
  fatherName: string;
  referenceName: string;
  locationName: string;
  email: string;
  dateOfBirth: string;
  dateOfJoining: string;
  appt: string;
  phoneNumber: string;
  selectCourse: string;
  courseDuration: string;
  gender: string;
  address: string;
  createdOn?: any;
}